requirements.txt
Django==5.0.6
requests==2.32.3
django-apscheduler==0.6.2
djangorestframework==3.15.2
reportlab==4.2.2
